requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
networkx==2.8.8
node2vec==0.4.6
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
streamlit==1.28.1
plotly==5.17.0
streamlit-option-menu==0.3.6
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.7.4
pypdf==3.17.1
pytest==7.4.3
httpx==0.25.2
faker==20.1.0
joblib==1.3.2
tqdm==4.66.1
gensim==4.3.2
scipy==1.15.3
